* Added capnp feature keyword   * Helps allow for Windows compilation. Capnp doesn't export symbols   for dlls and this is causing issues in supporting capnp on Windows   * capnp is featured out by default now * There are now two errors left in VS 2017 compilation. Both are related to the usage of madara::enable_if_same_decayed. In VS, this macro system is accidentally double-defining the macro for_each_field. I'll have to look more into this later or reimplement mfs and mabundle without Cereal or Capnp

+++ include/madara/knowledge/ThreadSafeContext.cpp
@@ -1907,8 +1907,8 @@ int64_t ThreadSafeContext::save_as_karl(
       char* result_copy = new char[settings.buffer_size];
       memcpy(result_copy, result.c_str(), result.size() + 1);
 
-      int size =
-          settings.encode(result_copy, result.size(), settings.buffer_size);
+      int size = settings.encode(
+        result_copy, (int)result.size(), (int)settings.buffer_size);
 
       if (size < 0)
       {
